Drop unnecessary dependency on IPy
IPy is only used to get the version of an IP (4 or 6), but the IPs come
from reasonably trustable sources and if incorrect would at worse make
the test fail -- which is valid behavior.

So, no need for a real library to distinguish them, do a naive check
based on the presence of version-specific delimiters.