Mails: don't cache configuration values too early
As the user configuration is now loaded after the mails module, there
is a race in the mail thread that fetches configuration options, which
may or may not already have the user's value.

So, remove caching of those value, which leads for them to be read much
later after everything has properly been initialized.  The timeout
setting might still have the default value on the first loop run, but
this doesn't matter as the server isn't started at this point anyway.